move: TDD 진행한 테스트 코드 디렉토리 이동
과제 테스트 통과를 위해 패키지 전체를 이동시켰습니다.
패키지를 이동하는 과정에서 뭔가 오류가 있었는지, Observer 클래스가 사라져서 다시 만들어서 옮겼습니다.
