--- src/test/java/racingcar/tdd/observer/Subject.java
@@ -1,7 +1,0 @@
-package racingcar.tdd.observer;
-
-public interface Subject {
-    void addObserver(Observer o);
-    void removeObserver(Observer o);
-    void notifyObservers();
-}
