--- src/test/java/racingcar/tdd/service/RaceCarEnroller.java
@@ -1,15 +1,0 @@
-package racingcar.tdd.service;
-
-import java.util.LinkedHashSet;
-import java.util.List;
-import racingcar.tdd.domain.Car;
-
-public class RaceCarEnroller {
-    private final LinkedHashSet<Car> cars = new LinkedHashSet<>();
-    public LinkedHashSet<Car> enrollCars(List<String> validatedCarNames) {
-        for (String validatedCarName : validatedCarNames) {
-            cars.add(new Car(validatedCarName));
-        }
-        return cars;
-    }
-}
