--- src/test/java/racingcar/tdd/validator/Validator.java
@@ -1,5 +1,0 @@
-package racingcar.tdd.validator;
-
-public interface Validator<T> {
-    T validate(T input);
-}
