--- src/test/java/racingcar/tdd/view/InputView.java
@@ -1,9 +1,0 @@
-package racingcar.tdd.view;
-
-import camp.nextstep.edu.missionutils.Console;
-
-public class InputView {
-    public static String getUserInput() {
-        return Console.readLine();
-    }
-}
