+++ src/test/java/tdd/observer/Observer.java
@@ -1,0 +1,7 @@
+package tdd.observer;
+
+import tdd.domain.Car;
+
+public interface Observer {
+    void display(Car car);
+}
