+++ src/test/java/tdd/observer/Subject.java
@@ -1,0 +1,7 @@
+package tdd.observer;
+
+public interface Subject {
+    void addObserver(Observer o);
+    void removeObserver(Observer o);
+    void notifyObservers();
+}
