+++ src/test/java/tdd/validator/Validator.java
@@ -1,0 +1,5 @@
+package tdd.validator;
+
+public interface Validator<T> {
+    T validate(T input);
+}
