+++ src/test/java/tdd/view/InputView.java
@@ -1,0 +1,9 @@
+package tdd.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputView {
+    public static String getUserInput() {
+        return Console.readLine();
+    }
+}
